Change size of neutral field in flattened_Mutation to avoid trailing padding.

+++ fwdpy11/src/_opaque_mutations.cc
@@ -29,7 +29,7 @@ struct flattened_Mutation
     double pos, s, h;
     fwdpp::uint_t g;
     decltype(fwdpy11::Mutation::xtra) label;
-    std::int8_t neutral;
+    std::int16_t neutral;
 };
 
 PYBIND11_MAKE_OPAQUE(std::vector<fwdpy11::Mutation>);
